requirements.txt
langchain
langchain-community
faiss-cpu
groq
streamlit
python-dotenv
PyMuPDF
sentence-transformers
